pyproject.toml
[tool.poetry]
name = "citibank-statement-downloader"
version = "0.1.0"
description = "Python library that downloads e-statements from 0"
repository = "https://github.com/benjamin-awd/citibank-statement-downloader"
authors = ["Benjamin Dornel <benjamindornel@gmail.com>"]
readme = "README.md"
packages = [
    { include = "citibank", from = "src" },
]
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
google-auth-oauthlib = "^1.2.0"
google-auth = "^2.26.0"
google-api-python-client = "^2.111.0"
pydantic-settings = "^2.1.0"
google-cloud-secret-manager = "^2.17.0"
selenium = "^4.16.0"
python-dateutil = "^2.8.2"
pydantic = "^2.5.3"
google-cloud-storage = "^2.14.0"
selenium-stealth = "^1.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pylint = "^3.0.3"
isort = "^5.12.0"
taskipy = "^1.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.4"
pytest-xdist = "^3.3.1"
pysnooper = "^1.2.0"
google-api-python-client-stubs = "^1.23.0"
mypy = "^1.6.1"
types-requests = "^2.31.0.20231231"
types-python-dateutil = "^2.8.19.14"
pytest-mock = "^3.12.0"
requests-mock = "^1.11.0"
ipykernel = "^6.29.5"

[tool.taskipy.tasks]
format = "isort . && black ."
lint = "flake8 src/citibank && pylint src/citibank"
full_test = "pytest -n auto"
mypy = "mypy src/citibank"
ci = "poetry run task format && poetry run task lint && poetry run task test"
test = "pytest ."

[tool.pylint]
good-names = ["i", "j", "k"]
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "no-member"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disable_error_code = [
    "annotation-unchecked",
]

[[tool.mypy.overrides]]
module = [
    "citibank.*",
    "google.*",
    "google_auth_oauthlib.flow",
]
ignore_missing_imports = true
